+++ appbuilder/tests/test_agent.py
@@ -64,18 +64,8 @@ class TestAgentRuntime(unittest.TestCase):
         for it in answer.content:
             self.assertIs(type(it), str)
 
-    def test_chainlit_agent_import_failed(self):
-        """ 测试import chainlit失败 """
-        component = AppBuilderClient(self.app_id)
-        agent = AgentRuntime(component=component)
-        subprocess.check_call([sys.executable, "-m", "pip", "uninstall", "-y", "chainlit"])
-        with self.assertRaises(ImportError):
-            agent.chainlit_agent()
-        subprocess.check_call([sys.executable, "-m", "pip", "install", "chainlit"])
-
     def test_chainlit_agent_component_error(self):
         """ 测试chainlit agent组件错误 """
-        subprocess.check_call([sys.executable, "-m", "pip", "install", "chainlit"])
         component = Component()
         agent = AgentRuntime(component=component)
         with self.assertRaises(ValueError):
